(update) organize and create new fuctions

+++ src/ft_check.c
@@ -29,11 +29,11 @@ int	ft_isdigit(char *str)
 
 
 
-int	ft_atoi(const char *str)
+size_t	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	size_t	i;
+	size_t	sign;
+	size_t	result;
 
 	result = 0;
 	i = 0;
